+++ src/program/GameLoop.h
@@ -86,6 +86,9 @@ signals:
     void getRamWatch(std::string &watch);
 
     void getTimeTrace(int type, unsigned long long hash, std::string stacktrace);
+    
+    /* Savestates have been invalidated by thread change */
+    void invalidateSavestates();
 };
 
 #endif