+++ src/program/ui/InputEditorModel.h
@@ -22,7 +22,6 @@
 
 #include <QtCore/QAbstractTableModel>
 #include <vector>
-#include <array>
 #include <stdint.h>
 
 #include "../Context.h"
@@ -138,15 +137,18 @@ public slots:
     /* Register a savestate for display */
     void registerSavestate(int slot, unsigned long long frame);
 
+    /* Invalidate all savestates */
+    void invalidateSavestates();
+
 private:
     Context *context;
     MovieFile *movie;
 
-    /* Array of framecount for savestates */
-    std::array<unsigned long long, 10> savestate_frames;
-
     /* Last saved/loaded state */
     unsigned long long last_savestate = 0;
+
+    /* Framecount of the last invalidation */
+    uint64_t invalid_frame = 0;
 
     /* Freeze the vertical scroll, used for rewind */
     bool freeze_scroll = false;