+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Basic MacOS support
 * Input editor: option to disable autoscroll (#182)
 * Input editor: option to seeks to current frame after rewind (#134)
+* Input editor: show state invalidation as gray area when threads changed
 
 ### Changed